Refactor to transfer get_location_time_zone method to base_repo
