+++ app/repositories/meal_session_repo.py
@@ -144,25 +144,6 @@ class MealSessionRepo(BaseRepo):
             return datetime(year=int(type_split[0]), month=int(type_split[1]), day=int(type_split[2]))
 
     @staticmethod
-    def get_location_time_zone(location_id):
-        """
-        Get the time zone of a particular location
-
-        :param location_id: string representing location id
-        :return: timezone object
-        :raises: AttributeError, pytz.exceptions.UnknownTimeZoneError
-        """
-        location_repo = LocationRepo()
-        location = location_repo.get(location_id)
-
-        try:
-            return pytz.timezone('Africa/' + location.name)
-        except AttributeError:
-            return AttributeError
-        except pytz.exceptions.UnknownTimeZoneError:
-            return pytz.exceptions.UnknownTimeZoneError
-
-    @staticmethod
     def validate_meal_session_already_exists(**kwargs):
         """
         Check whether a particular meal session already exists
